circle added

+++ Points/Source.cpp
@@ -4,7 +4,7 @@ int main() {
 	MyPoint p; 
 	std::cin >> p;
 	std::cout << p << std::endl;
-	int x = p;
+	int x = (int)p;
 	std::cout << x << std::endl;
 	system("pause");
 	return 0;
